Upstream's redundant-cast cleanup touched ftKb_SpecialN_800EED50, which this
branch had already moved out of ftkirby.c when the Kirby copy code was split
into its own translation unit. The conflicted region stays deleted from
ftkirby.c and the cast is dropped in ftkirbycopy.c instead, where the function
now lives.

Both Kirby translation units are unchanged by the merge: ftkirbyspeciallw is
44 of 44 symbols exact and ftkirbycopy is 96 of 97.

+++ src/melee/ft/chara/ftGameWatch/ftGw_SpecialHi.c
@@ -198,7 +198,7 @@ void ftGw_SpecialAirHi_IASA(HSD_GObj* gobj)
     float facing_dir;
     float stick_range_threshold;
 
-    if ((u32) fp->cmd_vars[0] == 0) {
+    if (fp->cmd_vars[0] == 0) {
         stick_x = stickGetDir(fp->input.lstick.x, 0.0f);
         if (stick_x > gawAttrs->x58_GAMEWATCH_RESCUE_STICK_RANGE) {
             stick_range_threshold =
