+++ src/melee/gm/gmhomerun.c
@@ -157,7 +157,7 @@ void gm_801B9A3C(GameScene* arg0)
                 vs->data.players[0].color, 1, gm_804D68F8);
     data->players[0].xD_b2 = true;
     gm_8016F088(data);
-    gm_80181A00(data->players[0].c_kind, (u8) data->players[0].xA);
+    gm_80181A00(data->players[0].c_kind, data->players[0].xA);
     gm_80180B18();
 }
