+++ src/melee/lb/lbbgflash.c
@@ -209,8 +209,8 @@ void fn_8001FEC4(HSD_GObj* gobj, s32 code)
     case 0:
         if ((u32) mode == 3U) {
             y = 0;
-            while (y <= (s32) data->x38) {
-                if (y == (s32) data->x38) {
+            while (y <= data->x38) {
+                if (y == data->x38) {
                     s32 width;
                     u8 strip_h;
                     s32 neg_y;
